Implemented the stratergy pattern
Changed the whole SessionController to three different timer controllers with an additional super class TimerController. I also changed the main so the programmer can switch between different implementations of the timers. The timers are in the beginning fase and it needs to be further developed on how they differ from each other and with the tools they can offer.